+++ examples/browser/index.js
@@ -1,0 +1,1 @@
+console.log(OpenBCIUtilities); // eslint-disable-line