+++ src/debug.js
@@ -1,7 +1,5 @@
-var Buffer = require('safe-buffer').Buffer;
-module.exports = {
-  debugBytes
-};
+
+import { Buffer } from 'buffer/';
 
 /**
  * @description Output passed bytes on the console as a hexdump, if enabled
@@ -9,7 +7,7 @@ module.exports = {
  * @param data - bytes to output, a buffer or string
  * @private
  */
-function debugBytes (prefix, data) {
+export default function debugBytes (prefix, data) {
   if (typeof data === 'string') data = new Buffer(data);
 
   console.log('Debug bytes:');